+++ doc/changelog.rst
@@ -12,6 +12,7 @@ Bug fixes
 2. Fixed typos and supported fields in docs (fixes :github:issue:`1105` and :github:issue:`1106`).
 3. Fixed bug where mixed ``jnt_limited`` joints were not being constrained correctly.
 4. Made ``device_put`` type validation more verbose (fixes :github:issue:`1113`).
+5. Removed empty EFC rows from `MJX`, for joints with no limits (fixes :github:issue:`1117`).
 
 Documentation
 ^^^^^^^^^^^^^